requirements.txt
Flask==1.1.1
Flask-DebugToolbar==0.11.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
python-dotenv==0.12.0
Scrapy==2.0.1